.github/workflows/files (20)/backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pytest
pytest-asyncio